Fix echo service
