+++ user/echo.go
@@ -10,16 +10,16 @@ import (
 
 func echoRequest(ur *url.URL) ([]byte, error) {
 	if ur.Scheme != "test" {
-		return nil, errors.Errorf("invalid test scheme")
+		return nil, errors.Errorf("invalid scheme for echo")
 	}
 	if ur.Host != "echo" {
-		return nil, errors.Errorf("invalid echo host")
+		return nil, errors.Errorf("invalid host for echo")
 	}
 
 	path := ur.Path
 	path = strings.TrimPrefix(path, "/")
 	paths := strings.Split(path, "/")
-	if len(paths) != 2 {
+	if len(paths) != 3 {
 		return nil, errors.Errorf("path invalid %s", path)
 	}
 	username := paths[0]
@@ -27,15 +27,20 @@ func echoRequest(ur *url.URL) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
+	msg := paths[2]
+	un, err := url.QueryUnescape(msg)
+	if err != nil {
+		return nil, err
+	}
+	msg = un
 
-	msg := ur.Query().Get("s")
-	user := &User{
+	usr := &User{
 		Service: "echo",
 		KID:     kid,
 		Name:    username,
 	}
 
-	if err := Verify(msg, user); err != nil {
+	if err := Verify(msg, usr); err != nil {
 		return nil, err
 	}
