+++ user/echo_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/keys-pub/keys"
 	"github.com/keys-pub/keys/docs"
+	"github.com/keys-pub/keys/link"
 	"github.com/keys-pub/keys/request"
 	"github.com/keys-pub/keys/tsutil"
 	"github.com/keys-pub/keys/user"
@@ -30,8 +31,8 @@ func TestResultEcho(t *testing.T) {
 	err = user.Verify(msg, usr)
 	require.NoError(t, err)
 
-	urs := "test://echo/alice/" + sk.ID().String() + "?s=" + url.QueryEscape(strings.ReplaceAll(msg, "\n", " "))
-	expected := `test://echo/alice/kex132yw8ht5p8cetl2jmvknewjawt9xwzdlrk2pyxlnwjyqrdq0dawqqph077?s=BEGIN+MESSAGE.+c0ypzQnuMjHRspp+4e0pl3TYCllN7ZG+MfKStJEnWVz5Uxt+lHFJtaTmEjPdy43+aOvtlDN9ZKwtQqS+WzHAKQB7RxKTCKq+6Xr2MZHgg4UNRDb+Zy2loGoGN3Mvxd4+r7FIwpZOJPE1JEq+D2gGjkgLByR9CFG+2aCgRgZZwl5UAa4+6bmBzjE1RqUnMN5+RDaVlacMSHyIP0d+IbCHwBmy0ZnY9pb+T0X.+END+MESSAGE.`
+	urs := "test://echo/alice/" + sk.ID().String() + "/" + url.QueryEscape(strings.ReplaceAll(msg, "\n", " "))
+	expected := `test://echo/alice/kex132yw8ht5p8cetl2jmvknewjawt9xwzdlrk2pyxlnwjyqrdq0dawqqph077/BEGIN+MESSAGE.+c0ypzQnuMjHRspp+4e0pl3TYCllN7ZG+MfKStJEnWVz5Uxt+lHFJtaTmEjPdy43+aOvtlDN9ZKwtQqS+WzHAKQB7RxKTCKq+6Xr2MZHgg4UNRDb+Zy2loGoGN3Mvxd4+r7FIwpZOJPE1JEq+D2gGjkgLByR9CFG+2aCgRgZZwl5UAa4+6bmBzjE1RqUnMN5+RDaVlacMSHyIP0d+IbCHwBmy0ZnY9pb+T0X.+END+MESSAGE.`
 	require.Equal(t, expected, urs)
 
 	sc := keys.NewSigchain(sk.ID())
@@ -68,4 +69,36 @@ func TestResultEcho(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, 1, len(kids))
 	require.Equal(t, keys.ID("kex132yw8ht5p8cetl2jmvknewjawt9xwzdlrk2pyxlnwjyqrdq0dawqqph077"), kids[0])
+}
+
+func TestRequestVerifyEcho(t *testing.T) {
+	sk := keys.NewEdX25519KeyFromSeed(testSeed(0x01))
+
+	clock := tsutil.NewTestClock()
+	req := request.NewMockRequestor()
+	ds := docs.NewMem()
+	scs := keys.NewSigchains(ds)
+	users := user.NewUsers(ds, scs, req, clock)
+
+	usrSign, err := user.NewForSigning(sk.ID(), "echo", "alice")
+	require.NoError(t, err)
+	msg, err := usrSign.Sign(sk)
+	require.NoError(t, err)
+	msg = url.QueryEscape(strings.ReplaceAll(msg, "\n", " "))
+
+	urs := "test://echo/alice/" + sk.ID().String() + "/" + msg
+
+	norm, err := link.Echo.NormalizeURLString("alice", urs)
+	require.NoError(t, err)
+
+	usr := &user.User{
+		KID:     sk.ID(),
+		Name:    "alice",
+		Service: "echo",
+		URL:     norm,
+	}
+
+	result := users.RequestVerify(context.TODO(), usr)
+	t.Logf("result: %+v", result)
+	require.Equal(t, user.StatusOK, result.Status)
 }
